chore: CE changes for Linting ModuleInputs
