+++ app/client/src/ce/plugins/Linting/lib/entity/types.ts
@@ -1,0 +1,18 @@
+import type { Diff } from "deep-diff";
+
+export enum ENTITY_TYPE {
+  ACTION = "ACTION",
+  WIDGET = "WIDGET",
+  APPSMITH = "APPSMITH",
+  JSACTION = "JSACTION",
+  PAGELIST = "PAGELIST",
+}
+
+export interface IEntity {
+  getName(): string;
+  getId(): string;
+  getType(): ENTITY_TYPE;
+  getRawEntity(): unknown;
+  getConfig(): unknown;
+  computeDifference(entity?: IEntity): Diff<unknown>[] | undefined;
+}
