+++ app/client/src/ee/plugins/Linting/lib/entity/entityConstructorMap.ts
@@ -1,0 +1,1 @@
+export * from "ce/plugins/Linting/lib/entity/entityConstructorMap";
