+++ app/client/src/ee/plugins/Linting/lib/entity/isDynamicEntity.ts
@@ -1,0 +1,1 @@
+export * from "ce/plugins/Linting/lib/entity/isDynamicEntity";
