+++ app/client/src/ee/plugins/Linting/lib/entity/types.ts
@@ -1,0 +1,1 @@
+export * from "ce/plugins/Linting/lib/entity/types";
